fix printf warnings and argument size errors which causes issues on x32 when debugging is enabled
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@36916 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ libmvutil/src/mailbox.c
@@ -200,7 +200,7 @@ int          mailbox_get_pollable_event(mailbox_t* mb)
         if (num < 0) 
             perrlog("write");
         if (num < mb->size) 
-            errlog(ERR_CRITICAL,"write truncated (%i/%i)\n",num,mb->size);
+            errlog(ERR_CRITICAL,"write truncated (%i/%i)\n",(int)num,(int)mb->size);
     }
 
     MB_UNLOCK(mb);
